+++ source/kernel/cpu.h
@@ -134,7 +134,7 @@ inline void CpuId::CheckIfValid() {
   }
 }
 
-inline uint8_t CpuId::GetApicId() {
+inline uint32_t CpuId::GetApicId() {
   return apic_ctrl->GetApicIdFromCpuId(*this);
 }
